adding components for consistent view for clinicDashboard  (#14)
* Navigated the KPT page for clinic staff

* Create DataCenter.js

Data Center UI

* Changing DataCenter.js location

Changing DataCenter.js to be under Manager folder and commenting it out since there is no use for it yet (waiting for iteration 2

* Update DataCenter.js

* added components for consistency

added two components for filter and tables for consistency with all the pages to have same visual.
and removed all the repeated tables with different styles from all the files.
also restructured the clinicDashboard.css and removed repeated names

---------

Co-authored-by: Md Ripon Hossain <ripon.du.chem@gmail.com>
Co-authored-by: Smilehead03 <saksham.03mahajan@gmail.com>