Final full iteration 1 Features (closing the branch) (#17)
* Getting patients from DB- in the appointment page

linking doctors and patients in the appointment page. it was creating new appointment without linking

* added patients list for clinic management

* Navigated the KPT page for clinic staff

* Navigated the KPT page for clinic staff

* Create DataCenter.js

Data Center UI

* Changing DataCenter.js location

Changing DataCenter.js to be under Manager folder and commenting it out since there is no use for it yet (waiting for iteration 2

* Update DataCenter.js

* added components for consistency

added two components for filter and tables for consistency with all the pages to have same visual.
and removed all the repeated tables with different styles from all the files.
also restructured the clinicDashboard.css and removed repeated names

* Improved code clarity by adding reason-based comments across appointment Firestore handlers; added author attribution (Vaibhav Kalia)

* Implement task creation, live task listener, and status updates using Firestore

* created to define the main application routing, protect routes by user role, and connect all staff, doctor, and manager dashboards under a single React Router setup.

* tested: update App test

* fixing mapping to Kpi view for doctor

* final css files update

---------

Co-authored-by: Md Ripon Hossain <ripon.du.chem@gmail.com>
Co-authored-by: Smilehead03 <saksham.03mahajan@gmail.com>
Co-authored-by: vaibhavkalia9585 <vaibhavkalia9585@gmail.com>